+++ library/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
+- Improve performance of `enum_` and `enumAsync` schema by caching values
+- Fix bug in `getDefaults`, `getDefaultsAsync`, `getFallbacks` and `getFallbacksAsync` for falsy but not `undefined` values (issue #356)
 - Change ISO timestamp regex to support timestamps with lower and higher millisecond accuracy (pull request #353)
 
 ## v0.25.0 (December 26, 2023)